dexter added , sound fixed, mqtt fixed , some bug fixes v 1.2.9

+++ app/src/main/java/com/akaf/nikoodriver/data/responses/offersHistoryResponse/CancelLocation.kt
@@ -1,6 +1,6 @@
 package com.akaf.nikoodriver.data.responses.offersHistoryResponse
 
 data class CancelLocation(
-    val coordinates: List<Int>,
+    val coordinates: List<Double>,
     val type: String
 )
